requirements.txt
pandas
numpy
scikit-learn
matplotlib
azure-ai-language-questionanswering
openai
pydantic
requests
python-dotenv
azure-ai-translation-text
streamlit